Endpoints done. Ready for frontend fun

+++ app/main_mod/main_mod_controllers.py
@@ -30,6 +30,9 @@ idError = {}
 idError['Error'] = 'ID Required'
 idErrorJson = json.dumps(idError)
 
+noResultError = {}
+noResultError['Error'] = 'No Result Found'
+noResultErrorJson = json.dumps(noResultError)
 '''
   Add objects
 '''
@@ -87,60 +90,72 @@ def editFeature():
   featureRequested = request.json
   if('id' not in featureRequested or featureRequested['id']==''):
     return idErrorJson
-  editFeature = db.session.query(Feature).filter_by(id=featureRequested['id']).one()
-  for field in featureRequested:
-    if field == 'title':
-      editFeature.title=featureRequested['title']
-    elif field == 'description':
-      editFeature.description=featureRequested['description']
-    elif field == 'client_id':
-      editFeature.client_id=featureRequested['client_id']
-    elif field == 'priority':
-      editFeature.priority=featureRequested['priority']
-    elif field == 'target_date':
-      editFeature.targetDate=datetime.datetime.strptime(
-        featureRequested['targetDate'], '%m/%d/%Y')
-    elif field == 'productArea_id':
-      editFeature.productArea_id=featureRequested['productArea_id']
-    elif field == 'submitter_id':
-      editFeature.submitter_id=featureRequested['submitter_id']
+  editFeature = db.session.query(Feature).filter_by(id=featureRequested['id']).first()
+  if(editFeature):
+    for field in featureRequested:
+      if field == 'title':
+        editFeature.title=featureRequested['title']
+      elif field == 'description':
+        editFeature.description=featureRequested['description']
+      elif field == 'client_id':
+        editFeature.client_id=featureRequested['client_id']
+      elif field == 'priority':
+        editFeature.priority=featureRequested['priority']
+      elif field == 'target_date':
+        editFeature.targetDate=datetime.datetime.strptime(
+          featureRequested['targetDate'], '%m/%d/%Y')
+      elif field == 'productArea_id':
+        editFeature.productArea_id=featureRequested['productArea_id']
+      elif field == 'submitter_id':
+        editFeature.submitter_id=featureRequested['submitter_id']
+    db.session.add(editFeature)
+    db.session.commit()
+    return redirect('/features')
 
-  db.session.add(editFeature)
-  db.session.commit()
-  return redirect('/features')
+  else:
+    return noResultErrorJson
 
 @mainModBase.route('/edit_client', methods=['POST'])
 def editClient():
   clientRequested = request.json
   if('id' not in clientRequested or clientRequested['id']==''):
     return idErrorJson
-  editClient = db.session.query(Client).filter_by(id=clientRequested['id']).one()
-  editClient.client_name = clientRequested['client_name']
-  db.session.add(editClient)
-  db.session.commit()
-  return redirect('/clients')
+  editClient = db.session.query(Client).filter_by(id=clientRequested['id']).first()
+  if(editClient):
+    editClient.client_name = clientRequested['client_name']
+    db.session.add(editClient)
+    db.session.commit()
+    return redirect('/clients')
+  else:
+    return noResultErrorJson
 
 @mainModBase.route('/edit_product_area', methods=['POST'])
 def editProductArea():
   productAreaRequested = request.json
   if('id' not in productAreaRequested or productAreaRequested['id']==''):
     return idErrorJson
-  editProductArea = db.session.query(ProductArea).filter_by(id=productAreaRequested['id']).one()
-  editProductArea.productArea_name = productAreaRequested['productArea_name']
-  db.session.add(editProductArea)
-  db.session.commit()
-  return redirect('/product_areas')
+  editProductArea = db.session.query(ProductArea).filter_by(id=productAreaRequested['id']).first()
+  if(editProductArea):
+    editProductArea.productArea_name = productAreaRequested['productArea_name']
+    db.session.add(editProductArea)
+    db.session.commit()
+    return redirect('/product_areas')
+  else:
+    return noResultErrorJson
   
 @mainModBase.route('/edit_user', methods=['POST'])
 def editUser():
   userRequested = request.json
   if('id' not in userRequested or userRequested['id']==''):
     return idErrorJson
-  editUser = db.session.query(User).filter_by(id=userRequested['id']).one()
-  editUser.user_name = userRequested['user_name']
-  db.session.add(editUser)
-  db.session.commit()
-  return redirect('/users')
+  editUser = db.session.query(User).filter_by(id=userRequested['id']).first()
+  if(editUser):
+    editUser.user_name = userRequested['user_name']
+    db.session.add(editUser)
+    db.session.commit()
+    return redirect('/users')
+  else:
+    return noResultErrorJson
 
 '''
   Delete objects
@@ -151,10 +166,13 @@ def deleteFeature():
   if('id' not in featureRequested or featureRequested['id']==''):
     return idErrorJson
   deleteFeature = db.session.query(Feature).filter_by(
-    id=featureRequested['id']).one()
-  db.session.delete(deleteFeature)
-  db.session.commit()
-  return redirect('/features')
+    id=featureRequested['id']).first()
+  if(deleteFeature):
+    db.session.delete(deleteFeature)
+    db.session.commit()
+    return redirect('/features')
+  else:
+    return noResultErrorJson
 
 @mainModBase.route('/delete_client', methods=['POST'])
 def deleteClient():
@@ -162,15 +180,18 @@ def deleteClient():
   if('id' not in clientRequested or clientRequested['id']==''):
     return idErrorJson
   deleteClient = db.session.query(Client).filter_by(
-    id=clientRequested['id']).one()
-  relatedFeatures = db.session.query(Feature).filter_by(
-    client_id=clientRequested['id']).all()
-  for feature in relatedFeatures:
-    feature.client_id = None
-    db.session.add(feature)
-  db.session.delete(deleteClient)
-  db.session.commit()
-  return redirect('/clients')
+    id=clientRequested['id']).first()
+  if(deleteClient):
+    relatedFeatures = db.session.query(Feature).filter_by(
+      client_id=clientRequested['id']).all()
+    for feature in relatedFeatures:
+      feature.client_id = None
+      db.session.add(feature)
+    db.session.delete(deleteClient)
+    db.session.commit()
+    return redirect('/clients')
+  else:
+    return noResultErrorJson
 
 @mainModBase.route('/delete_product_area', methods=['POST'])
 def deleteProductArea():
@@ -178,15 +199,18 @@ def deleteProductArea():
   if('id' not in productAreaRequested or productAreaRequested['id']==''):
     return idErrorJson
   deleteProductArea = db.session.query(ProductArea).filter_by(
-    id=productAreaRequested['id']).one()
-  relatedFeatures = db.session.query(Feature).filter_by(
-    productArea_id=productAreaRequested['id']).all()
-  for feature in relatedFeatures:
-    feature.productArea_id = None
-    db.session.add(feature)
-  db.session.delete(deleteProductArea)
-  db.session.commit()
-  return redirect('/product_areas')
+    id=productAreaRequested['id']).first()
+  if(deleteProductArea):
+    relatedFeatures = db.session.query(Feature).filter_by(
+      productArea_id=productAreaRequested['id']).all()
+    for feature in relatedFeatures:
+      feature.productArea_id = None
+      db.session.add(feature)
+    db.session.delete(deleteProductArea)
+    db.session.commit()
+    return redirect('/product_areas')
+  else:
+    return noResultErrorJson
   
 @mainModBase.route('/delete_user', methods=['POST'])
 def deleteUser():
@@ -194,12 +218,15 @@ def deleteUser():
   if('id' not in userRequested or userRequested['id']==''):
     return idErrorJson
   deleteUser = db.session.query(User).filter_by(
-    id=userRequested['id']).one()
-  relatedFeatures = db.session.query(Feature).filter_by(
-    submitter_id=userRequested['id']).all()
-  for feature in relatedFeatures:
-    feature.submitter_id = None
-    db.session.add(feature)
-  db.session.delete(deleteUser)
-  db.session.commit()
-  return redirect('/users')
+    id=userRequested['id']).first()
+  if(deleteUser):
+    relatedFeatures = db.session.query(Feature).filter_by(
+      submitter_id=userRequested['id']).all()
+    for feature in relatedFeatures:
+      feature.submitter_id = None
+      db.session.add(feature)
+    db.session.delete(deleteUser)
+    db.session.commit()
+    return redirect('/users')
+  else:
+    return noResultErrorJson
